+++ src/pages/_app.js
@@ -1,5 +1,10 @@
-import '@/styles/globals.css'
+import Layout from '@/components/ui/Layout';
+import '@/styles/globals.css';
 
 export default function App({ Component, pageProps }) {
-  return <Component {...pageProps} />
+    return (
+        <Layout>
+            <Component {...pageProps} />
+        </Layout>
+    );
 }
